🚀 Release 2.3.0 (#484)
* refs #53279 error during indexing products

* fix php cs

* Delete 202/docker/run_for_unittest_v8.sh

* Delete 202/.phpunit.result.cache

* Update CartCarrierAssociation.php

* Update phpstan.neon

* Update RulesManager.php

* Update backword-compatibility.php

* Update CartCarrierAssociation.php

* Update RulesManager.php

* refs #52542 phpstan

* phpstan

* phpcsfixer

* phpstan version

* phpstan error

* module version 2.2.0

* refs #51957 move the cloud sync form to the top

* module key

* refs #47192 amazon marketplace. intraco_vat_number (#471)

* Feature partial refund (51016) #460

* refs #51016 new option in BO

* refs #51016 partial refund

* refs #51016 small correction

* refs #51016 refund shipping

* php-sc-fixer

* refs #51016 handling failed tickets

* module version. phpcsfixer. phpstan

* Add prefix to cdiscount email (54264) #472 

* refs #54264 add prefix to cdiscount email

* phpstan

* refs #53069 colissimo relay id (#474)

* Feature/53940 import order discounts (#475)

* refs #53940 import the marketplace discounts

* phpcsfixer

* phpstan

* phpstan

* phpcsfixer

* phpcsfixer

* refs #54472 wrong tax rate

* phpstan

* refs #53940 unit-test

* refs #54191 send attachment title (#473)

* refs #54191 send attachment title

* phpcsfixer

* small correction

---------


* refs #53940 translations (#476)

* refs #53940 translations

* small correction

* refs #54481 sending mpn as mpn_sf

* refs #47192 small correction

* Hotfix/psvalidator (#478)

* phpcsfixer

* psvalidator

* phpstan

* refs #54195 compatibility with ps9 (#479)

* refs #54481 rename mpn into mpn_sf (#480)

* refs #54713 order shipping weight (#482)

---------

+++ classes/actions/ShoppingfeedProductSyncStockActions.php
@@ -42,13 +42,13 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
     public function prepareBatch()
     {
         $this->conveyor['preparedBatch'] = [];
+        /** @var Shoppingfeed $sfModule */
         $sfModule = Module::getInstanceByName('shoppingfeed');
         $token = new ShoppingfeedToken($this->conveyor['id_token']);
 
         /** @var ShoppingfeedProduct $sfProduct */
         foreach ($this->conveyor['batch'] as $sfProduct) {
             $sfReference = $sfModule->mapReference($sfProduct);
-
             // The developer can skip products to sync by overriding
             // ShoppingFeed::mapReference and have it return false
             if (empty($sfReference)) {
@@ -81,16 +81,16 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
     {
         $shoppingfeedApi = ShoppingfeedApi::getInstanceByToken($this->conveyor['id_token']);
         if ($shoppingfeedApi == false) {
-            ProcessLoggerHandler::logError(
+            ProcessLoggerHandler::logError( // @phpstan-ignore-line
                 static::getLogPrefix($this->conveyor['id_token']) . ' ' .
                     $this->l('Could not retrieve Shopping Feed API.', 'ShoppingfeedProductSyncStockActions'),
                 'Product'
             );
-            Registry::increment('errors');
+            Registry::increment('errors'); // @phpstan-ignore-line
 
             return false;
         }
-        $limit = Configuration::getGlobalValue(Shoppingfeed::STOCK_SYNC_MAX_PRODUCTS);
+        $limit = (int) Configuration::getGlobalValue(Shoppingfeed::STOCK_SYNC_MAX_PRODUCTS);
         $preparedBatch = $this->conveyor['preparedBatch'];
         foreach (array_chunk($preparedBatch, $limit, true) as $products) {
             $res = $shoppingfeedApi->updateMainStoreInventory($products, $this->conveyor['shoppingfeed_store_id']);
@@ -106,7 +106,7 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
                 $sfProduct = $products[$reference]['sfProduct'];
 
                 if (false === Validate::isLoadedObject($sfProduct)) {
-                    ProcessLoggerHandler::logError(
+                    ProcessLoggerHandler::logError( // @phpstan-ignore-line
                         sprintf(
                             static::getLogPrefix($this->conveyor['id_token']) . ' ' .
                             $this->l('Cannot retrieve a product for a reference %s', 'ShoppingfeedProductSyncStockActions'),
@@ -118,7 +118,7 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
                     continue;
                 }
 
-                ProcessLoggerHandler::logInfo(
+                ProcessLoggerHandler::logInfo( // @phpstan-ignore-line
                     sprintf(
                         static::getLogPrefix($this->conveyor['id_token']) . ' ' .
                             $this->l('Updated %s qty: %s', 'ShoppingfeedProductSyncStockActions'),
@@ -129,7 +129,7 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
                     $sfProduct->id_product
                 );
 
-                Registry::increment('updatedProducts');
+                Registry::increment('updatedProducts'); // @phpstan-ignore-line
 
                 unset($products[$reference]);
 
@@ -142,7 +142,7 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
             foreach ($products as $data) {
                 $sfProduct = $data['sfProduct'];
 
-                ProcessLoggerHandler::logInfo(
+                ProcessLoggerHandler::logInfo( // @phpstan-ignore-line
                     sprintf(
                         static::getLogPrefix($this->conveyor['id_token']) . ' ' .
                             $this->l('%s not referenced in the Shopping Feed catalog or unmodified Shopping Feed stock - qty: %s', 'ShoppingfeedProductSyncStockActions'),
@@ -152,7 +152,7 @@ class ShoppingfeedProductSyncStockActions extends ShoppingfeedProductSyncActions
                     'Product',
                     $sfProduct->id_product
                 );
-                Registry::increment('not-in-catalog');
+                Registry::increment('not-in-catalog'); // @phpstan-ignore-line
 
                 $sfProduct->delete();
             }
